Added makefile; Added READMEs

+++ scripts/create-rpm-package.sh
@@ -24,20 +24,26 @@ export GOOS=linux
 
 go build -o reporter ../.
 
-# RPM
+# RPM PACKAGE
 RPMPACKAGE=go-mongo-reporter-${VERSION}
 mkdir -p $RPMPACKAGE
-
 cp reporter $RPMPACKAGE
 cp ../reporter.service $RPMPACKAGE
-
 tar -cvzf $RPMPACKAGE.tar.gz $RPMPACKAGE/
 
+# RPMBUILD
 rpmdev-setuptree
 cp $RPMPACKAGE.tar.gz ~/rpmbuild/SOURCES
 rpmbuild -ba ../build/rpm/reporter.spec
+if [ $? -ne 0 ]; then { echo "Build failed, aborting." ; exit 1; } fi
 
 cp ~/rpmbuild/RPMS/x86_64/$RPMPACKAGE-$BUILD_NUMBER.el9.x86_64.rpm ../build/generated-packages
+# CLEANUP
 rm reporter
 rm -rf $RPMPACKAGE 
 rm $RPMPACKAGE.tar.gz
+
+cd ../build/generated-packages
+echo ""
+echo "Build succesful, new RPM package created to: $PWD"
+echo ""
